pyproject.toml
[tool.poetry]
name = "orionx-api-client"
version = "1.1.0"
description = "API Client for Orionx Exchange"
authors = ["Ignacio Guerrero <ignacio.tolosa@usach.cl>"]
license = "MIT"
readme = "README.md"
packages = [{include = "orionx_api_client"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4"
requests = "^2.31.0"
graphql-core = "^3.2.3"
gql = "^3.4.1"
requests-toolbelt = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
pytest-recording = "^0.13.0"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"
pytest-mock = "^3.11.1"
autoflake = "^2.2.1"

[tool.isort]
profile = "black"

[tool.pylint.format]
max-line-length = "88"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
